% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{attr.range}
\alias{attr.range}
\title{=========================================================================#
Compute denominator of the diff formula
for each attribute x (column) in my.mat, max(x) - min(x)}
\usage{
attr.range(my.mat)
}
\arguments{
\item{my.mat}{attribute matrix}
}
\value{
Numeric vectors representing the denominators in the diff formula
}
\description{
=========================================================================#
Compute denominator of the diff formula
for each attribute x (column) in my.mat, max(x) - min(x)
}
