% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{case.control.3sets}
\alias{case.control.3sets}
\title{case-control interaction effect data}
\format{
Data frame with 300 samples (100 samples in train/holdout/test) and 
100 features with moderate (interaction) effect size.
}
\usage{
data(case.control.3sets)
}
\description{
case-control interaction effect data
}
\keyword{datasets}
