% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkPackages}
\alias{checkPackages}
\title{checkPackages}
\usage{
checkPackages(pkg)
}
\arguments{
\item{pkg}{character vector of package names to install if not already.}
}
\value{
null There is nothing returned.
}
\description{
Utility to check for package installation
}
\examples{
packages <- c("ggplot2", "CORElearn", "reshape2", "dplyr", "pROC", "plotROC")
checkPackages(packages)  
}
