% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation2.R
\name{createMainEffects}
\alias{createMainEffects}
\title{Create a simulated data set with main effects}
\usage{
createMainEffects(
  n.e = 1000,
  n.db = 70,
  n.ns = 30,
  label = "class",
  pct.imbalance = 0.5,
  sv.db = c("A", "B"),
  sv.ns = c("A", "B"),
  sd.b = 1,
  sd.gam = 1,
  sd.u = 1,
  conf = FALSE,
  distr.db = NA,
  p.b = 0.3,
  p.gam = 0.3,
  p.ov = p.b/2
)
}
\arguments{
\item{n.e}{number of variables}

\item{n.db}{sample size in database}

\item{n.ns}{sample size in newsample}

\item{label}{A character vector for the name of the outcome column. class for classification
and qtrait for regression}

\item{pct.imbalance}{A numeric percentage to indicate proportion of the imbalaced samples. 
0 means all controls and 1 mean all cases.}

\item{sv.db}{batches}

\item{sv.ns}{batches}

\item{sd.b}{a numeric for sd.b?}

\item{sd.gam}{a numeric for sd.gam?}

\item{sd.u}{a numeric for sd.u?}

\item{conf}{a flag for conf?}

\item{distr.db}{a numeric for distr.db?}

\item{p.b}{a numeric for p.b?}

\item{p.gam}{a numeric for p.gam?}

\item{p.ov}{a numeric for p.ov?}
}
\value{
A list with:
\describe{
  \item{db}{database creation variables}
  \item{vars}{variables used in simulation}
}
}
\description{
\eqn{X = BS + \Gamma G + U}
}
\details{
S = Biological group                                                                                                                   m
G = Batch
U = random error

NOTE:  If you use conf=TRUE, then you must have exactly two surrogate
variables in the database this function only allows for confounding in
the database, not confounding in the new samples
}
\references{
Leek,J.T. and Storey,J.D. (2007) Capturing heterogeneity in gene expression
studies by surrogate variable analysis. PLoS Genet., 3, 1724–1735
}
\seealso{
Other simulation: 
\code{\link{splitDataset}()}
}
\concept{simulation}
