% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detectionStats}
\alias{detectionStats}
\title{detectionStats}
\usage{
detectionStats(functional, positives, num_all_vars = NULL)
}
\arguments{
\item{functional}{character vector of functional/true attribute names.}

\item{positives}{character vector of attribute names of positive associations (null hypothesis rejected or some threshold).}

\item{num_all_vars}{total number of variables, needed for TN calculation for MCC}
}
\value{
list with elements TP, FP, FN, TPR, FPR, precision, recall, F1, MCC, and summary message (string).
}
\description{
Given a vector functional (true) attribute names and a vector of positive
association attribute names, returns detection statistics like recall and precision.
Note you must include the number of variables if you want to compute TN and MCC.
}
\examples{
functional <- c("var1", "var2", "var3")
detected <- c("var1", "var2", "var4")
detectionStats(functional, detected)
}
