% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbixGAIN.R
\name{dmgain}
\alias{dmgain}
\title{Differential modularity genetic association network algorithm (dmGAIN) algorithm.}
\usage{
dmgain(labelledDataFrame)
}
\arguments{
\item{labelledDataFrame}{\code{data.frame} with samples in rows, variables in columns
and classLabel in the last column.}
}
\value{
results \code{matrix} of variable by variable differential modularity values.
}
\description{
\code{dmgain}
}
\examples{
data(testdata10)
rinbixDmgain <- dmgain(testdata10)
}
\references{
\itemize{
  \item \url{https://github.com/hexhead/inbix}{C++ inbix on github}
}
}
\seealso{
Other GAIN functions: \code{\link{dcgain}},
  \code{\link{fitInteractionModel}},
  \code{\link{fitMainEffectModel}},
  \code{\link{gainToSimpleSIF}},
  \code{\link{getInteractionEffects}},
  \code{\link{getMainEffects}},
  \code{\link{regainParallel}}, \code{\link{regain}}
}
\concept{GAIN functions}
\concept{inbix synonym functions
\code{\link{dcgain}} for differential correlation GAIN.}
\keyword{array}
\keyword{models}
