% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regain.R
\name{EpistasisKatz}
\alias{EpistasisKatz}
\title{Rank variables by EpistasisKatz algorithm.}
\usage{
EpistasisKatz(A = NULL, alpha = NULL, beta = NULL, magnitude.sort = T)
}
\arguments{
\item{A}{\code{matrix} network of features either in adjacency or interaction format.}

\item{alpha}{\code{numeric} a vector with numeric values.}

\item{beta}{\code{numeric} either a vector of constant values or prior knowledge.}

\item{magnitude.sort}{default (T) is to rank variable scores by their magnitude
so negative effects are also ranked. Using False will force negative effects to be
ranked last.}
}
\value{
features ranking with features name.
}
\description{
This function can be use as original Katz algorithm or as EpistasisKatz
that incorporates prior knowledge.
}
\details{
\code{EpistasisKatz}
}
