% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regain.R
\name{fitInteractionModel}
\alias{fitInteractionModel}
\title{\code{fitInteractionModel}}
\usage{
fitInteractionModel(
  labelledDataFrame,
  variableIndices,
  depVarName,
  regressionFamily,
  numCovariates,
  excludeMainEffects
)
}
\arguments{
\item{labelledDataFrame}{\code{labelledDataFrame.frame} with variables in columns and samples in rows.}

\item{variableIndices}{\code{vector} of column indices of variable pairs.}

\item{depVarName}{\code{string} name of the classLabel variable.}

\item{regressionFamily}{\code{string} glm regression family name.}

\item{numCovariates}{\code{numeric}  of covariates included.}

\item{excludeMainEffects}{\code{logical} indicating whether to exclude main effect terms.}
}
\value{
\code{data.frame} with variable, convergence status, beta coefficient,
p-value, standard error and standardized beta columns.
}
\description{
\code{fitInteractionModel}
}
