% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbixGAIN.R
\name{gainToSimpleSIF}
\alias{gainToSimpleSIF}
\title{GAIN matrix to simple interaction format (SIF) converter.}
\usage{
gainToSimpleSIF(gainMatrix)
}
\arguments{
\item{gainMatrix}{\code{matrix} GAIN matrix.}
}
\value{
sifDF \code{data.frame} of node1 weight node2.
}
\description{
\code{gainToSimpleSIF}
}
\examples{
data(testdata100ME4)
rinbixRegain <- regainParallel(testdata100ME4, stdBetas = TRUE, absBetas = TRUE)
gainSIF <- gainToSimpleSIF(rinbixRegain)
}
\seealso{
Other GAIN functions: \code{\link{dcgain}},
  \code{\link{dmgain}}, \code{\link{fitInteractionModel}},
  \code{\link{fitMainEffectModel}},
  \code{\link{getInteractionEffects}},
  \code{\link{getMainEffects}},
  \code{\link{regainParallel}}, \code{\link{regain}}
}
\concept{GAIN functions}
\keyword{array}
