% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation2.R
\name{generate_structured_corrmat}
\alias{generate_structured_corrmat}
\title{generate_structured_corrmat}
\usage{
generate_structured_corrmat(g = NULL, num.variables = 100,
  hi.cor.tmp = 0.8, lo.cor.tmp = 0.2, graph.type = "Erdos-Renyi",
  plot.graph = F, make.diff.cors = F, nbias = 1, use.Rcpp = F)
}
\arguments{
\item{g}{random graph}

\item{num.variables}{number of independent variables in data matrix}

\item{graph.type}{either Erdos-Renyi or Scale-Free graph}

\item{plot.graph}{logical indicating whether to plot graph or not}

\item{make.diff.cors}{logical indicating whether case correlation matrix for differential correlation is being created or not}

\item{nbias}{number of functional interaction variables}

\item{use.Rcpp}{if true use Rcpp to correct negative eigenvalues}

\item{hi.cor}{upper baseline pairwise functional correlation in control group}

\item{lo.cor}{lower baseline pairwise functional correlation in control group}
}
\value{
A list containing:
\describe{
  \item{cor.mat}{structured correlation matrix}
  \item{deg.vec}{degree vector corresponding to network adjacency}
  \item{A.mat}{adjacency matrix corresponding to network}
  \item{sig.vars}{indices of functional variables}
}
}
\description{
parameters:
}
\examples{

}
