% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regain.R
\name{getInteractionEffects}
\alias{getInteractionEffects}
\title{\code{getInteractionEffects}}
\usage{
getInteractionEffects(
  dats,
  depVarName,
  reg.fam = "binomial",
  nCovs = 0,
  excludeMainEffects = F,
  indVarNames
)
}
\arguments{
\item{dats}{\code{labelledDataFrame.frame} with variables in columns and samples in rows.}

\item{depVarName}{\code{string} name of the classLabel variable.}

\item{reg.fam}{\code{string} glm regression family name.}

\item{nCovs}{\code{numeric}  of covariates included.}

\item{excludeMainEffects}{\code{logical} indicating whether to exclude main effect terms.}

\item{indVarNames}{Vector of independent variable names, to name output matrices}
}
\value{
\code{data.frame} with variable, convergence status, beta coefficient,
p-value, standard error and standardized beta columns.
}
\description{
\code{getInteractionEffects}
}
