% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regain.R
\name{getMainEffects}
\alias{getMainEffects}
\title{\code{getMainEffects}
Wrapper for fitMainEffectModel with sapply over independent variable names. 
Does not include covariates, so if you have covariates use 
indVarNames = colnames(dats)[1:(ncol(dats) - nCovs - 1)] 
if the columns are organized as: independent vars, covars, pheno}
\usage{
getMainEffects(
  dats,
  indVarNames = NULL,
  depVarName = NULL,
  reg.fam = "binomial",
  nCovs = 0
)
}
\arguments{
\item{dats}{\code{data.frame} with variables in columns and samples in rows.}

\item{indVarNames}{vector of independent variable names to loop over.}

\item{depVarName}{\code{string} name of the classLabel variable.}

\item{nCovs}{\code{numeric}  of included covariates.}

\item{regressionFamily}{\code{string} glm regression family name.}
}
\value{
mainEffectValues \code{vector} of main effect values.
}
\description{
\code{getMainEffects}
Wrapper for fitMainEffectModel with sapply over independent variable names. 
Does not include covariates, so if you have covariates use 
indVarNames = colnames(dats)[1:(ncol(dats) - nCovs - 1)] 
if the columns are organized as: independent vars, covars, pheno
}
