% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmSTIR.R
\name{glmSTIR}
\alias{glmSTIR}
\title{glmSTIR}
\usage{
glmSTIR(outcome, data.set, covars = "none", regression.type = "lm",
  nbd.method = "multisurf", nbd.metric = "manhattan",
  attr.diff.type = "numeric-abs", covar.diff.type = "numeric-abs",
  k = 0, sd.frac = 0.5, fdr.method = "bonferroni")
}
\arguments{
\item{outcome}{character name or length-m numeric outcome vector for linear regression, factor for logistic regression}

\item{data.set}{m x p matrix of m instances and p attributes, May also include outcome vector but then outcome should be name. Include attr names as colnames.}

\item{regression.type}{(\code{"lm"} or \code{"glm"})}

\item{nbd.method}{neighborhood method [\code{"multisurf"} or \code{"surf"} (no k) or \code{"relieff"} (specify k)]. Used by nearestNeighbors().}

\item{nbd.metric}{used in stirDistances for distance matrix between instances, default: \code{"manhattan"} (numeric). Used by nearestNeighbors().}

\item{attr.diff.type}{diff type for attributes (\code{"manhattan"} or \code{"euclidean"} for numeric). Phenotype diff same as attr.diff.type when lm regression. For glm, phenotype diff is match/mismatch.}

\item{covar.diff.type}{diff type for covariates (\code{"manhattan"} or \code{"euclidean"} for numeric).}

\item{k}{number of constant nearest hits/misses for \code{"relieff"} (fixed-k). Used by nearestNeighbors().
The default k=0 means use the expected SURF theoretical k with sd.frac (.5 by default)}

\item{sd.frac}{multiplier of the standard deviation from the mean distances; subtracted from mean for SURF or multiSURF.
The multiSURF default is sd.frac=0.5: mean - sd/2. Used by nearestNeighbors().}

\item{fdr.method}{for p.adjust (\code{"fdr"}, \code{"bonferroni"}, ...)}

\item{covar}{vector of names or indices of covariates in data.set for covariate correction. Or separate data matrix of covariates.}

\item{neighbor.pairs.idx}{nearest hit/miss matrix (2 columns), output from \code{find.neighbors}}
}
\value{
glmSTIR.stats.df: glmSTIR regression coefficients and p-values for each attribute
}
\description{
generalized linear model (GLM) based STatistical Inference Relief (STIR)
Computes Relief-based attribute statistical signficance for case/control or quantitative outcomes.
Allows for categorical (SNP) or numeric (expession) predictor data types. 
Allows for covariate correction.
}
\examples{
# Data interface options.
# Specify name ("qtrait") of outcome and data.set, which is a data frame including the outcome column.
# ReliefF fixed-k neighborhood, uses surf theoretical default (with sd.frac=.5) if you do not specify k or let k=0
restir.results.df <- glmSTIR("qtrait", train.data, regression.type="lm", nbd.method="relieff", nbd.metric = "manhattan", attr.diff.type="manhattan", covar.diff.type="manhattan", sd.frac=0.5, fdr.method="bonferroni")

# Specify column index (101) of outcome and data.set, which is a data frame including the outcome column.
restir.results.df <- glmSTIR(101, train.data, regression.type="lm", nbd.method="relieff", nbd.metric = "manhattan", attr.diff.type="manhattan", covar.diff.type="manhattan", k=10, fdr.method="bonferroni")

# if outcome vector (pheno.vec) is separate from attribute matrix
# multisurf
restir.results.df <- glmSTIR(pheno.vec, predictors.mat, regression.type="lm", nbd.method="multisurf", nbd.metric = "manhattan", attr.diff.type="manhattan", covar.diff.type="manhattan", sd.frac=0.5, fdr.method="bonferroni")
restir.positives <- row.names(restir.results.df[restir.results.df[,1]<.05,]) # reSTIR p.adj<.05
}
