% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hamming.binary}
\alias{hamming.binary}
\title{Hamming distance for a binary matrix
https://johanndejong.wordpress.com/2015/10/02/faster-hamming-distance-in-r-2/}
\usage{
hamming.binary(X)
}
\arguments{
\item{X}{Original matrix.}
}
\value{
Distance matrix with the Hamming metric.
}
\description{
Hamming distance for a binary matrix
https://johanndejong.wordpress.com/2015/10/02/faster-hamming-distance-in-r-2/
}
