% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{knnSURF.balanced}
\alias{knnSURF.balanced}
\title{knnSURF.balanced}
\usage{
knnSURF.balanced(class.vec, sd.frac = 0.5)
}
\arguments{
\item{class.vec}{vector of class labels to determine the minimum class size.}

\item{sd.frac}{fraction of the standard deviation from the mean of all pairwise distances, dead-band. The default value used by the SURF and multiSURF algorithms is 1/2.}
}
\value{
knn Theoretical number of neighbors.
}
\description{
Theoretical value for the number of expected neighbors for SURF or multiSURF 
(fixed or adaptive radius) neighborhoods, but adjusted for imbalanced data. 
We use our theoretical formula with twice the size of the minority class 
as the input sample size.
}
\examples{
k.surf.bal <- knnSURF.balanced(class.vec, .5)
}
