% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{knnSURF}
\alias{knnSURF}
\title{knnSURF}
\usage{
knnSURF(m.samples, sd.frac = 0.5)
}
\arguments{
\item{m.samples}{number of samples in data.}

\item{sd.frac}{fraction of the standard deviation from the mean of all pairwise distances, dead-band. The default value used by the SURF and multiSURF algorithms is 1/2.}
}
\value{
knn Number of neighbors.
}
\description{
Theoretical value for the number of expected neighbors for SURF or multiSURF
}
\examples{
k.surf <- knnSURF(200,.5)
}
