% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestNeighbors.R
\name{knnVec}
\alias{knnVec}
\title{knnVec}
\usage{
knnVec(neighbor.pairs.mat)
}
\arguments{
\item{neighbor.pairs.mat}{two columns of redundant "i,j" pairs from nearestNeighbors function}
}
\value{
knn.vec vector number of nearest neighbors for each instance
}
\description{
Number of neighbors for each sample (vector) from a neighbor-pair matrix.
}
\examples{
neighbor.pairs.idx <- nearestNeighbors(
  predictors.mat,
  nbd.method = "multisurf",
  nbd.metric = "manhattan",
  sd.frac = 0.5
)
mean(knnVec(neighbor.pairs.idx)) # average number of neighbors
}
