% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mdd.RNAseq.small}
\alias{mdd.RNAseq.small}
\title{RNA-Seq from the paper referenced below.}
\format{
Data frame with 157 mdd and hc subjects with 
500 randomly selected genes from the original 5912 genes.
}
\usage{
data(mdd.RNAseq.small)
}
\description{
RNA-Seq from the paper referenced below.
}
\keyword{datasets}
