% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npdr.R
\name{npdr}
\alias{npdr}
\title{npdr}
\usage{
npdr(outcome, dataset, regression.type = "binomial",
  attr.diff.type = "numeric-abs", nbd.method = "multisurf",
  nbd.metric = "manhattan", knn = 0, msurf.sd.frac = 0.5,
  covars = "none", covar.diff.type = "match-mismatch",
  glmnet.alpha = 1, glmnet.lower = 0, glmnet.family = "binomial",
  rm.attr.from.dist = c(), neighbor.sampling = NULL,
  padj.method = "bonferroni", verbose = FALSE)
}
\arguments{
\item{outcome}{character name or length-m numeric outcome vector for linear regression, factor for logistic regression}

\item{dataset}{m x p matrix of m instances and p attributes, May also include outcome vector but then outcome should be name. Include attr names as colnames.}

\item{regression.type}{(\code{"lm"} or \code{"binomial"})}

\item{attr.diff.type}{diff type for attributes (\code{"numeric-abs"} or \code{"numeric-sqr"} for numeric, \code{"allele-sharing"} or \code{"match-mismatch"} for SNP). Phenotype diff uses same numeric diff as attr.diff.type when lm regression. For glm-binomial, phenotype diff is \code{"match-mismatch"}.}

\item{nbd.method}{neighborhood method [\code{"multisurf"} or \code{"surf"} (no k) or \code{"relieff"} (specify k)]. Used by nearestNeighbors().}

\item{nbd.metric}{used in npdrDistances for distance matrix between instances, default: \code{"manhattan"} (numeric). Used by nearestNeighbors().}

\item{knn}{number of constant nearest hits/misses for \code{"relieff"} (fixed-k). Used by nearestNeighbors().
The default knn=0 means use the expected SURF theoretical k with msurf.sd.frac (.5 by default)}

\item{msurf.sd.frac}{multiplier of the standard deviation from the mean distances; subtracted from mean for SURF or multiSURF.
The multiSURF default is msurf.sd.frac=0.5: mean - sd/2. Used by nearestNeighbors().}

\item{covars}{optional vector or matrix of covariate columns for correction. Or separate data matrix of covariates.}

\item{covar.diff.type}{string (or string vector) specifying diff type(s) for covariate(s) (\code{"numeric-abs"} for numeric or \code{"match-mismatch"} for categorical).}

\item{glmnet.alpha}{penalty mixture for npdrNET: default alpha=1 (lasso, L1) alpha=0 (ridge, L2)}

\item{glmnet.lower}{lower limit for coefficients for npdrNET: lower.limits=0 npdrNET default}

\item{rm.attr.from.dist}{attributes for removal (possible confounders) from the distance matrix calculation. Argument for nearestNeighbors. None by default c().}

\item{neighbor.sampling}{NULL or \code{"unique"} if you want to use only unique neighbor pairs (used in nearestNeighbors)}

\item{padj.method}{for p.adjust (\code{"fdr"}, \code{"bonferroni"}, ...)}

\item{glment.family}{"binomial" for logistic regression, "gaussian" for regression}
}
\value{
npdr.stats.df: npdr fdr-corrected p-value for each attribute ($pval.adj [1]), raw p-value ($pval.attr [2]), and regression coefficient (beta.attr [3])
}
\description{
Nearest-Neighbor Projected-Distance Regression (npdr) 
generalized linear model (GLM) extension of STatistical Inference Relief (STIR)
Computes attribute statistical signficance with logistic for case/control and linear model for quantitative outcomes.
NPDR allows for categorical (SNP) or numeric (expession) predictor data types. 
NPDR allows for covariate correction.
Observations in the model are projected-distance differences between neighbors.
}
\examples{
# Data interface options.
# Specify name ("qtrait") of outcome and dataset, which is a data frame including the outcome column.
# ReliefF fixed-k neighborhood, uses surf theoretical default (with msurf.sd.frac=.5) if you do not specify k or let k=0
npdr.results.df <- npdr("qtrait", train.data, regression.type="lm", nbd.method="relieff", nbd.metric = "manhattan", attr.diff.type="manhattan", covar.diff.type="manhattan", msurf.sd.frac=0.5, padj.method="bonferroni")

# Specify column index (101) of outcome and dataset, which is a data frame including the outcome column.
npdr.results.df <- npdr(101, train.data, regression.type="lm", nbd.method="relieff", nbd.metric = "manhattan", attr.diff.type="manhattan", covar.diff.type="manhattan", knn=10, padj.method="bonferroni")

# if outcome vector (pheno.vec) is separate from attribute matrix
# multisurf
npdr.results.df <- npdr(pheno.vec, predictors.mat, regression.type="lm", nbd.method="multisurf", nbd.metric = "manhattan", attr.diff.type="manhattan", covar.diff.type="manhattan", msurf.sd.frac=0.5, padj.method="bonferroni")
# attributes with npdr adjusted p-value less than .05 
npdr.positives <- row.names(npdr.results.df[npdr.results.df$pva.adj<.05,]) # npdr p.adj<.05
}
