% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestNeighbors.R
\name{npdrDiff}
\alias{npdrDiff}
\title{npdrDiff}
\usage{
npdrDiff(
  a,
  b,
  diff.type = c("manhattan", "numeric-abs", "numeric-sqr", "allele-sharing",
    "match-mismatch", "correlation-data"),
  norm.fac = 1
)
}
\arguments{
\item{a}{value of attribute for first instance. Vector for correlation-data.}

\item{b}{value of attribute for second instance. Vector for correlation-data.}

\item{diff.type}{Metric for the difference computation.}

\item{norm.fac}{Normalization factor.}
}
\value{
Value or vector of differences between two vectors element-wise.
}
\description{
A diff is a function that computes the difference of values for an attribute between two instances.
It is used for attribute selection for attribute diffs and phenotype diffs.
This function is vectorized: input a and b can be two vectors of values for one attribute.
}
