% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npdrLearner.R
\name{npdrDistances2}
\alias{npdrDistances2}
\title{npdrDistances2}
\usage{
npdrDistances2(attr.mat1, attr.mat2, metric = "manhattan")
}
\arguments{
\item{attr.mat1}{m1 x p matrix of m instances and p attributes}

\item{attr.mat2}{m2 x p matrix of m instances and p attributes}

\item{metric}{for distance matrix between instances 
(default: \code{"manhattan"}, others include \code{"euclidean"},
and for GWAS \code{"allele-sharing-manhattan"}).}
}
\value{
matrix of m1 instances x m2 instances pairwise distances.
}
\description{
Create m1 x m2 distance matrix between two datasets,
(m1 instances and p attributes in dataset1 and
m2 instances and p attributes in dataset2).
Datasets should not include phenotype column.
Uses function dist2 from flexclust.
Used by nearestNeighbors2().
}
\examples{
train_dat <- case.control.3sets$train
valid_dat <- case.control.3sets$validation
dist.mat <- npdrDistances2(
  train_dat[, names(train_dat) != "class"], 
  valid_dat[, names(valid_dat) != "class"], 
  metric = "manhattan"
)
}
