% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{predictors.mat}
\alias{predictors.mat}
\title{case-control interaction effect data, features only}
\format{
Data frame with 100 samples (100 samples in train/holdout/test) and 
100 features/predictors with moderate (main) effect size.
No "class" column.
}
\usage{
data(predictors.mat)
}
\description{
case-control interaction effect data, features only
}
\keyword{datasets}
