% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qtrait.3sets}
\alias{qtrait.3sets}
\title{quantitative trait main effect data}
\format{
Data frame with 300 samples (100 samples in train/holdout/test) and 
100 features with moderate (main) effect size.
}
\usage{
data(qtrait.3sets)
}
\description{
quantitative trait main effect data
}
\keyword{datasets}
