% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbixGAIN.R
\name{regainParallel}
\alias{regainParallel}
\title{Parallel execution of the regression genetic association interaction network (reGAIN) algorithm.}
\usage{
regainParallel(
  depVarName = "class",
  labelledDataFrame,
  stdBetas = FALSE,
  absBetas = FALSE,
  numCores = 2,
  verbose = FALSE,
  writeBetas = FALSE,
  interactOutput = "Betas",
  regressionFamily = "binomial"
)
}
\arguments{
\item{depVarName}{\code{string} name of class column (e.g., "class")}

\item{labelledDataFrame}{\code{data.frame} with variable in columns and samples in rows;
the last column should be labeled 'Class' and be 0 or 1 values.}

\item{stdBetas}{\code{logical} to use standardized beta coefficients.}

\item{absBetas}{\code{logical} to use absolute value of beta coefficients.}

\item{numCores}{\code{numeric} number of processor cores to use in mclapply}

\item{verbose}{\code{logical} to send verbose messages to stdout.}

\item{writeBetas}{\code{logical} indicating whether to write beta values to separate file.}

\item{interactOutput}{\code{string} output in return matrix, "indicating betas "Betas", "stdBetas", "Pvals".}

\item{regressionFamily}{\code{string} glm regression family name.}
}
\value{
regainMatrix \code{matrix} of variable by variable regression coefficients.
}
\description{
\code{regainParallel}
}
\examples{
data(testdata10)
rinbixRegain <- regainParallel(testdata10, stdBetas = TRUE, absBetas = TRUE)
}
\references{
\itemize{
  \item \url{http://www.nature.com/gene/journal/v11/n8/full/gene201037a.html}
  {Genes & Immunity Paper}
  \item \url{https://github.com/hexhead/inbix}{C++ inbix on github}
}
}
\seealso{
Other GAIN functions: 
\code{\link{dcgain}()},
\code{\link{dmgain}()},
\code{\link{fitInteractionModel}()},
\code{\link{fitMainEffectModel}()},
\code{\link{gainToSimpleSIF}()},
\code{\link{getInteractionEffects}()},
\code{\link{getMainEffects}()},
\code{\link{regain}()}

Other inbix synonym functions: 
\code{\link{dcgain}()},
\code{\link{getMainEffects}()},
\code{\link{regain}()}
}
\concept{GAIN functions}
\concept{inbix synonym functions}
\keyword{array}
\keyword{models}
\keyword{regression}
