% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reliefDetected}
\alias{reliefDetected}
\title{reliefDetected}
\usage{
reliefDetected(results.df, functional, top.pct)
}
\arguments{
\item{results.df}{dataframe of relief-sorted (high to low) attribute names from CORElearn}

\item{functional}{character vector of functional/true attribute names}

\item{p}{percentile of top relief scores compared with the functional list}
}
\value{
True positive rate: number of true postives divided by the number of functional
}
\description{
Given a vector functional (true) attribute names, a vector of sorted attribute names, and percentile 
threshold, returns true positive rate.
}
\examples{
functional.vars <- dataset$signal.names  
relief <- CORElearn::attrEval(as.factor(class) ~ ., data = dats, 
                              estimator = "ReliefFequalK",
                              costMatrix = NULL,
                              outputNumericSplits=FALSE,
                              kNearestEqual = floor(knnSURF(nrow(dats),.5)/2)) # fn from npdr
relief.order <- order(relief, decreasing = T)
relief.df <- data.frame(att=names(relief)[relief.order], rrelief=relief[relief.order])
reliefDetected(relief.df,functional.vars,p=.1)
}
