% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rfDetected}
\alias{rfDetected}
\title{rfDetected}
\usage{
rfDetected(results.df, functional, top.pct)
}
\arguments{
\item{results.df}{dataframe of sorted (high to low) attribute names from randomForest}

\item{functional}{character vector of functional/true attribute names}

\item{p}{percentile of top relief scores compared with the functional list}
}
\value{
True positive rate: number of true postives divided by the number of functional
}
\description{
Given a vector functional (true) attribute names, a vector of sorted attribute names, and percentile 
threshold, returns true positive rate.
}
\examples{
functional.vars <- dataset$signal.names  
ranfor.fit <- randomForest(as.factor(class) ~ ., data = dats) 
rf.importance <- importance(ranfor.fit)  
rf.sorted<-sort(rf.importance, decreasing=T, index.return=T)
rf.df <-data.frame(att=rownames(rf.importance)[rf.sorted$ix],rf.scores=rf.sorted$x)
rfDetected(rf.df,functional.vars,p=.1)
}
