% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rownames2columns}
\alias{rownames2columns}
\title{Convert rownames to column.}
\usage{
rownames2columns(df, var = "rowname")
}
\arguments{
\item{df}{Input dataframe.}

\item{var}{Name of new column.}
}
\value{
A data frame with a new column of row names.
}
\description{
Adapted from https://github.com/tidyverse/tibble/blob/516449bbb0c76925b93b703b68d7979a53d7cdee/R/rownames.R.
}
