% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation2.R
\name{splitDataset}
\alias{splitDataset}
\title{Split a data set for machine learning classification}
\usage{
splitDataset(
  all.data = NULL,
  pct.train = 0.5,
  pct.holdout = 0.5,
  pct.validation = 0,
  label = "class"
)
}
\arguments{
\item{all.data}{A data frame of n rows by d colums of data plus a label column}

\item{pct.train}{A numeric percentage of samples to use for traning}

\item{pct.holdout}{A numeric percentage of samples to use for holdout}

\item{pct.validation}{A numeric percentage of samples to use for testing}

\item{label}{A character vector of the data column name for the outcome label. class for classification
and qtrait for regression.}
}
\value{
A list containing:
\describe{
  \item{train}{traing data set}
  \item{holdout}{holdout data set}
  \item{validation}{validation data set}
}
}
\description{
Return data.sets as a list of training set, holdout set and validation set
according to the predefined percentage of each partition
default is a 50-50 split into training and holdout, no testing set
code class/label/phenotypes as 1 and -1.
User can manage the simulation data to be dichotomious/quantitative using label (class/qtrait)
}
\examples{
data.sets <- splitDataset(case.control.3sets$train, label = "class")

}
\seealso{
Other simulation: 
\code{\link{createMainEffects}()}
}
\concept{simulation}
