% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testdata10}
\alias{testdata10}
\title{Test data from Rinbix}
\format{
An object of class \code{data.frame} with 10 rows and 11 columns.
}
\usage{
data(testdata10)
}
\description{
Test data from Rinbix
}
\keyword{datasets}
