% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestNeighbors.R
\name{uniqueNeighbors}
\alias{uniqueNeighbors}
\title{uniqueNeighbors}
\usage{
uniqueNeighbors(neighbor.pairs.idx)
}
\arguments{
\item{neighbor.pairs.idx}{two columns of redundant "i,j" pairs from nearestNeighbors function}
}
\value{
unique.idx vector index of the unique pairs from the input neighborhood
}
\description{
Find pairs of unique nearest neighbors pairs from possible redundant pairs. 
Used as options (neighbor.sampling="unique") in nearestNeighbors and npdr functions.
}
\examples{
unique.neighbor.pairs.idx <- uniqueNeighbors(neighbor.pairs.idx)  # unique neighbor pairs

}
