% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uniReg}
\alias{uniReg}
\title{uniReg}
\usage{
uniReg(outcome, dataset, regression.type = "lm", padj.method = "fdr",
  covars = "none")
}
\arguments{
\item{outcome}{string with name of class column.}

\item{dataset}{data matrix with predictor columns and outcome column.}

\item{regression.type}{"lm" or "binomial"}

\item{padj.method}{for p.adjust (\code{"fdr"}, \code{"bonferroni"}, ...)}

\item{covars}{optional vector or matrix of covariate columns for correction. Or separate data matrix of covariates.}
}
\value{
matrix of beta, p-value and adjusted p-value, sorted by p-value.
}
\description{
Univariate logistic or linear regression for a dataset.
}
\examples{
lr.results <- uniReg(outcome="class", dataset=case.control.data, regression.type="binomial")
}
