% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{univariateLogisticRegression}
\alias{univariateLogisticRegression}
\title{univariateLogisticRegression}
\usage{
univariateLogisticRegression(outcome, dataset)
}
\arguments{
\item{outcome}{string with name of class column.}

\item{dataset}{data matrix with predictor columns and outcome column.}
}
\value{
matrix of beta, p-value and adjusted p-value, sorted by p-value.
}
\description{
Univariate logistic regression for a data set.
}
\examples{
lr.results <- univariateLogisticRegression(outcome="class", dataset=case.control.data)
}
