% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{univarRegress}
\alias{univarRegress}
\title{univarRegress}
\usage{
univarRegress(outcome, dataset, regression.type = "lm",
  covars = "none")
}
\arguments{
\item{outcome}{string with name of class column.}

\item{dataset}{data matrix with predictor columns and outcome column.}
}
\value{
matrix of beta, p-value and adjusted p-value, sorted by p-value.
}
\description{
Univariate logistic or linear regression for a dataset.
}
\examples{
lr.results <- univarRegress(outcome="class", dataset=case.control.data, regression.type="glm")
}
