% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vwak}
\alias{vwak}
\title{vwak}
\usage{
vwak(dats = NULL, k.grid = NULL, verbose = F,
  attr.diff.type = "numeric-abs", corr.attr.names = NULL,
  separate.hitmiss.nbds = FALSE, label = "class")
}
\arguments{
\item{dats}{m x (p+1) data set of m instances and p attributes with 1 binary outcome or m x [p(p - 1) + 1] with p(p-1) correlations and 1 outcome. Outcome is last column for standard m x (p + 1) and first column for m x [p(p - 1) + 1] (no good reason for the difference).}

\item{k.grid}{increasing sequence of k values used as looping index. Default is seq(1,(nrow(dats)-1),by=1).}

\item{verbose}{logical indicating whether to print progress with loop. Default is FALSE, but TRUE also does not give anything useful.}

\item{attr.diff.type}{character indicating the type of attribute diff to use. Default is 'numeric-abs' for standard continuous data. Use 'correlation-data' for rs-fMRI data.}

\item{corr.attr.names}{character indicating names of ROIs for attr.diff.type='correlation-data'. Default is NULL.}

\item{separate.hitmiss.nbds}{logical indicating whether to compute hit/miss neighborhoods separately. Default is FALSE.}

\item{label}{character indicating type of response. Default is "class" and should not change as of yet.}
}
\value{
A list with:
\describe{
  \item{beta.mat}{p x k matrix of beta coefficients from NPDR}
  \item{pval.mat}{p x k matrix of P values corresponding to beta coefficients from NPDR}
}
}
\description{
Variable-Wise Adaptive k
method for optimizing NPDR scores for each attribute as a function of k
Computes p x k beta and P value matrices for a data set with p attributes
}
\examples{
# Main effect simulation for standard m x p data set
num.samples <- 100
num.variables <- 100
pct.imbalance <- 0.5
pct.signals <- 0.1
main.bias <- 0.5
sim.type <- "mainEffect"

dataset <- createSimulation2(num.samples=num.samples,
                             num.variables=num.variables,
                             pct.imbalance=pct.imbalance,
                             pct.signals=pct.signals,
                             main.bias=main.bias,
                             label="class",
                             sim.type=sim.type,
                             pct.mixed=0.5,
                             pct.train=0.5,
                             pct.holdout=0.5,
                             pct.validation=0,
                             verbose=T,
                             data.type="continuous")
dats <- rbind(dataset$train, dataset$holdout, dataset$validation)
dats <- dats[order(dats[,ncol(dats)]),]

# run Variable-Wise Adaptive k function
out <- vwak(dats=dats,
            k.grid=NULL,
            verbose=T,
            attr.diff.type="numeric-abs",
            label="class")
}
