% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation2.R
\name{createSimulation2}
\alias{createSimulation2}
\title{createSimulation2}
\usage{
createSimulation2(num.samples = 100, num.variables = 100,
  pct.imbalance = 0.5, pct.signals = 0.1, main.bias = 0.4,
  interaction.bias = 0.4, hi.cor = 0.8, lo.cor = 0.2,
  label = "class", sim.type = "mainEffect", pct.train = 0.5,
  pct.holdout = 0.5, pct.validation = 0, save.file = NULL,
  mix.type = NULL, pct.mixed = 0.5, verbose = FALSE,
  plot.graph = F, use.Rcpp = F, prob.connected = NULL,
  out.degree = NULL)
}
\arguments{
\item{num.samples}{number of samples}

\item{num.variables}{number of variables (features)}

\item{pct.imbalance}{fraction of num.samples that are cases}

\item{pct.signals}{fraction of num.variables that are functional}

\item{main.bias}{approximate effect size for main effect simulations}

\item{interaction.bias}{approximate effect size for interaction effects}

\item{hi.cor}{parameter to use for network-connected pairwise correlations}

\item{lo.cor}{parameter to use for network-non-connected pairwise correlations}

\item{label}{should just be "class" for binary response}

\item{sim.type}{a character that determines the type of simulation:
mainEffect/mainEffect_Erdos-Renyi/mainEffect_Scalefree/interactionErdos/interactionScalefree/mixed}

\item{pct.train}{fraction of num.samples used for training}

\item{pct.holdout}{fraction of num.samples used for holdout}

\item{pct.validation}{fraction of num.samples used for validation}

\item{save.file}{logical but not currently being used}

\item{mix.type}{character that determines the type of mixed effects simulation:
main-interactionErdos/main-interactionScalefree}

\item{pct.mixed}{fraction of functional variables with interaction effects}

\item{verbose}{logical indicating whether to display time required to generate simulation}

\item{plot.graph}{logical indicating whether to plot networks}

\item{use.Rcpp}{if true use Rcpp to correct negative eigenvalues}

\item{prob.connected}{probability of drawing an edge between two arbitrary vertices in Erdos-Renyi graph}

\item{out.degree}{out-degree of vertices in Scale-free graph}
}
\value{
A list with:
\describe{
  \item{train}{traing data set}
  \item{holdout}{holdout data set}
  \item{validation}{validation data set}
  \item{label}{the class label/column name}
  \item{signal.names}{the variable names with simulated signals}
  \item{elapsed}{total elapsed time}
  \item{A.mat}{adjacency matrix for random network (for interaction/mixed/main plus correlation)}
  \item{main.vars}{indices of main effect variables (for mixed simulations)}
  \item{int.vars}{indices of interaction effect variables (for mixed simulations)}
}
}
\description{
createSimulation2
}
\examples{
num.samples <- 100
num.variables <- 10

dataset <- createSimulation2(num.samples=num.samples,
                            num.variables=num.variables,
                            pct.imbalance=0.5,
                            pct.signals=0.2,
                            main.bias=0.4,
                            interaction.bias=1,
                            hi.cor=10,
                            lo.cor=0.2,
                            mix.type="main-interactionErdos",
                            label="class",
                            sim.type="mainEffect_Erdos-Renyi",
                            pct.mixed=0.5,
                            pct.train=0.5,
                            pct.holdout=0.5,
                            pct.validation=0,
                            plot.graph=F,
                            verbose=T)
}
