% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detectionStats}
\alias{detectionStats}
\title{detectionStats}
\usage{
detectionStats(functional, positives)
}
\arguments{
\item{functional}{character vector of functional/true attribute names.}

\item{positives}{character vector of attribute names of positive associations (null hypothesis rejected or some threshold).}
}
\value{
list with elements TP, FP, FN, TPR, FPR, precision, recall and summary message (string).
}
\description{
Given a vector functional (true) attribute names and a vector of positive
association attribute names, function returns detection statistics like recall and precision.
}
\examples{
functional <- case.control.3sets$signal.names  
positives <- row.names(glm.stir.results.df[glm.stir.results.df[,1]<.05,]) # p.adj<.05
glm.stir.detect.stats <- detectionStats(functional.case.control, glm.stir.positives)
cat(glm.stir.detect.stats$summary.msg)on(outcome="class", dataset=case.control.data)
}
