% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbixGAIN.R
\name{fitInteractionModel}
\alias{fitInteractionModel}
\title{Get the interaction effect of a pair of variables using generalized linear regression.}
\usage{
fitInteractionModel(labelledDataFrame, variableIndices, depVarName,
  regressionFamily, numCovariates, excludeMainEffects)
}
\arguments{
\item{labelledDataFrame}{\code{labelledDataFrame.frame} with variables in columns and samples in rows.}

\item{variableIndices}{\code{vector} of column indices of variable pairs.}

\item{depVarName}{\code{string} name of the classLabel variable.}

\item{regressionFamily}{\code{string} glm regression family name.}

\item{numCovariates}{\code{numeric}  of covariates included.}

\item{excludeMainEffects}{\code{logical} indicating whether to exclude main effect terms.}
}
\value{
\code{data.frame} with variable, convergence status, beta coefficient,
p-value, standard error and standardized beta columns.
}
\description{
\code{fitInteractionModel}
}
\seealso{
Other GAIN functions: \code{\link{dcgain}},
  \code{\link{dmgain}}, \code{\link{fitMainEffectModel}},
  \code{\link{gainToSimpleSIF}},
  \code{\link{getInteractionEffects}},
  \code{\link{getMainEffects}},
  \code{\link{regainParallel}}, \code{\link{regain}}
}
\concept{GAIN functions}
\keyword{array}
\keyword{internal}
\keyword{models}
\keyword{regression}
