% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geneLowVarianceFilter}
\alias{geneLowVarianceFilter}
\title{geneLowVarianceFilter}
\usage{
geneLowVarianceFilter(dataMatrix, percentile = 0.5)
}
\arguments{
\item{dataMatrix}{data matrix with predictors only, sample x gene}

\item{percentile}{percentile of low variance removed}
}
\value{
mask and filtered data
}
\description{
Low variance mask and filtered data for gene expression matrix.
}
