% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation2.R
\name{generate_structured_corrmat}
\alias{generate_structured_corrmat}
\title{generate_structured_corrmat}
\usage{
generate_structured_corrmat(
  g = NULL,
  num.variables = 100,
  hi.cor.tmp = 0.8,
  lo.cor.tmp = 0.2,
  hi.cor.fixed = 0.8,
  graph.type = "Erdos-Renyi",
  plot.graph = FALSE,
  make.diff.cors = FALSE,
  nbias = 1,
  use.Rcpp = FALSE
)
}
\arguments{
\item{g}{random graph}

\item{num.variables}{number of independent variables in data matrix}

\item{hi.cor.tmp}{high correlation proportion}

\item{lo.cor.tmp}{low correlation proportion}

\item{hi.cor.fixed}{high correlation for network connected/non-functional features}

\item{graph.type}{either Erdos-Renyi or Scale-Free graph}

\item{plot.graph}{logical indicating whether to plot graph or not}

\item{make.diff.cors}{logical indicating whether case correlation matrix for differential correlation is being created or not}

\item{nbias}{number of functional interaction variables}

\item{use.Rcpp}{logical, whether to use Rcpp or not}

\item{hi.cor}{upper baseline pairwise functional correlation in control group}

\item{lo.cor}{lower baseline pairwise functional correlation in control group}
}
\value{
A list containing:
\describe{
  \item{cor.mat}{structured correlation matrix}
  \item{deg.vec}{degree vector corresponding to network adjacency}
  \item{A.mat}{adjacency matrix corresponding to network}
  \item{sig.vars}{indices of functional variables}
}
}
\description{
parameters:
}
