% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbixGAIN.R
\name{getInteractionEffects}
\alias{getInteractionEffects}
\title{Get interaction effects from generalized linear model regression.}
\usage{
getInteractionEffects(depVarName = "class", labelledDataFrame,
  regressionFamily = "binomial", numCovariates = 0,
  writeBetas = FALSE, excludeMainEffects = FALSE,
  interactOutput = "Pvals", transformMethod = "", numCores = 2,
  verbose = FALSE)
}
\arguments{
\item{depVarName}{\code{string} name of class column (e.g., "class")}

\item{labelledDataFrame}{\code{data.frame} with variables in columns and samples in rows.}

\item{regressionFamily}{\code{string} glm regression family name.}

\item{numCovariates}{\code{numeric}  of included covariates.}

\item{writeBetas}{\code{logical} indicating whether to write beta values to separate file.}

\item{excludeMainEffects}{\code{logical} indicating whether to exclude main effect terms.}

\item{interactOutput}{\code{string} output in return matrix, "indicating betas "Betas", "stdBetas", "Pvals".}

\item{transformMethod}{\code{string} optional transform method.}

\item{numCores}{\code{numeric} number of processor cores to use in mclapply}

\item{verbose}{\code{logical} to send verbose messages to stdout.}
}
\value{
results \code{matrix} of variable by variable regression coefficients.
}
\description{
\code{getInteractionEffects}
}
\seealso{
Other GAIN functions: \code{\link{dcgain}},
  \code{\link{dmgain}}, \code{\link{fitInteractionModel}},
  \code{\link{fitMainEffectModel}},
  \code{\link{gainToSimpleSIF}},
  \code{\link{getMainEffects}},
  \code{\link{regainParallel}}, \code{\link{regain}}
}
\concept{GAIN functions}
\keyword{array}
\keyword{internal}
\keyword{models}
\keyword{regression}
