% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestNeighbors.R
\name{nearestNeighbors}
\alias{nearestNeighbors}
\title{nearestNeighbors}
\usage{
nearestNeighbors(attr.mat, nb.method = "multisurf",
  nb.metric = "manhattan", sd.vec = NULL, sd.frac = 0.5, k = 0,
  attr_removal_vec_from_dist_calc = c())
}
\arguments{
\item{attr.mat}{m x p matrix of m instances and p attributes}

\item{nb.method}{neighborhood method [\code{"multisurf"} or \code{"surf"} (no k) or \code{"relieff"} (specify k)]}

\item{nb.metric}{used in npdrDistances for distance matrix between instances, default: \code{"manhattan"} (numeric)}

\item{sd.frac}{multiplier of the standard deviation from the mean distances, subtracted from mean distance to create for SURF or multiSURF radius. The multiSURF default "dead-band radius" is sd.frac=0.5: mean - sd/2}

\item{k}{number of constant nearest hits/misses for \code{"relieff"} (fixed k). 
The default k=0 means use the expected SURF theoretical k with sd.frac (.5 by default) for relieff nbd.}

\item{attr_removal_vec_from_dist_calc}{attributes for removal (possible confounders) from the distance matrix calculation.}
}
\value{
Ri_NN.idxmat, matrix of Ri's (first column) and their NN's (second column)
}
\description{
Find nearest neighbors of each instance using relief.method
Used for regression stir (reSTIR) (no hits or misses specified in function).
Also used in reSTIR for case/control, but hit/miss is used in reSTIR function.
}
\examples{
# multisurf neighborhood with sigma/2 (sd.frac=0.5) "dead-band" boundary
neighbor.pairs.idx <- nearestNeighbors(predictors.mat, nb.method="multisurf", nb.metric = "manhattan", sd.frac = 0.5)
# reliefF (fixed-k) neighborhood using default k equal to theoretical surf expected value
# One can change the theoretical value by changing sd.frac (default 0.5)
neighbor.pairs.idx <- nearestNeighbors(predictors.mat, nb.method="relieff", nb.metric = "manhattan")
# reliefF (fixed-k) neighborhood with a user-specified k
neighbor.pairs.idx <- nearestNeighbors(predictors.mat, nb.method="relieff", nb.metric = "manhattan", k=10)

}
