% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestNeighbors.R
\name{npdrDiff}
\alias{npdrDiff}
\title{npdrDiff}
\usage{
npdrDiff(a, b, diff.type = "numeric-abs", norm.fac = 1)
}
\arguments{
\item{a}{value of attribute for first instance. Vector for correlation-data.}

\item{b}{value of attribute for second instance. Vector for correlation-data.}

\item{type}{diff rule for the given attribute data type, such as numeric, categorical or correlation-data vector.}
}
\value{
val diff or vector of diffs
}
\description{
A diff is a function that computes the difference of values for an attribute between two instances.
It is used for attribute selection for attribute diffs and phenotype diffs.
This function is vectorized: input a and b can be two vectors of values for one attribute.
}
