% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regain.R
\name{regain}
\alias{regain}
\title{\code{getMainEffects}}
\usage{
regain(
  dats,
  depVarName,
  indVarNames,
  reg.fam = "binomial",
  nCovs = 0,
  excludeMainEffects = F
)
}
\arguments{
\item{dats}{\code{labelledDataFrame.frame} with variables in columns and samples in rows.}

\item{depVarName}{\code{string} name of the classLabel variable.}

\item{indVarNames}{Vector of independent variable names, to name output matrices}

\item{reg.fam}{\code{string} glm regression family name.}

\item{nCovs}{\code{numeric}  of covariates included.}

\item{excludeMainEffects}{\code{logical} indicating whether to exclude main effect terms.}
}
\value{
regainMatrix \code{matrix} of variable by variable regression coefficients.
}
\description{
\code{getMainEffects}
}
