% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestNeighbors.R
\name{uniqueNeighbors}
\alias{uniqueNeighbors}
\title{uniqueNeighbors}
\usage{
uniqueNeighbors(neighbor.pairs)
}
\arguments{
\item{neighbor.pairs}{two columns of (possibly redundant) "i,j" pairs from nearestNeighbors function}
}
\value{
new neighborhood pair matrix of only unique pairs
unique neighbor pairs
}
\description{
Find pairs of unique nearest neighbors pairs from possible redundant pairs.
Used as options (neighbor.sampling="unique") in nearestNeighbors and npdr functions.
}
\examples{
neighbor.pairs.idx <- nearestNeighbors(
  predictors.mat, 
  nbd.method = "multisurf", 
  nbd.metric = "manhattan", 
  sd.frac = 0.5
)
head(uniqueNeighbors(neighbor.pairs.idx))

}
